PA2

+++ PA2/b07901021_pa2/src/MaxPlanarSubset.h
@@ -28,5 +28,5 @@ class ChordSet{
 
 };
 
-void traverse(int, int, ChordSet*, vector<int>*, short**);
+void traverse(int, int, ChordSet*, vector<int>*, int**);
 #endif
